Import 2.1.42pre1

+++ include/asm-sparc/byteorder.h
@@ -1,24 +1,24 @@
-multiline_comment|/* $Id: byteorder.h,v 1.12 1996/12/19 08:08:20 davem Exp $ */
+multiline_comment|/* $Id: byteorder.h,v 1.13 1997/05/26 23:37:46 davem Exp $ */
 macro_line|#ifndef _SPARC_BYTEORDER_H
 DECL|macro|_SPARC_BYTEORDER_H
 mdefine_line|#define _SPARC_BYTEORDER_H
 DECL|macro|ntohl
-mdefine_line|#define ntohl(x) (x)
+mdefine_line|#define ntohl(x) ((unsigned long int)(x))
 DECL|macro|ntohs
-mdefine_line|#define ntohs(x) (x)
+mdefine_line|#define ntohs(x) ((unsigned short int)(x))
 DECL|macro|htonl
-mdefine_line|#define htonl(x) (x)
+mdefine_line|#define htonl(x) ((unsigned long int)(x))
 DECL|macro|htons
-mdefine_line|#define htons(x) (x)
+mdefine_line|#define htons(x) ((unsigned short int)(x))
 multiline_comment|/* Some programs depend upon these being around. */
 DECL|macro|__constant_ntohl
-mdefine_line|#define __constant_ntohl(x) (x)
+mdefine_line|#define __constant_ntohl(x) ((unsigned long int)(x))
 DECL|macro|__constant_ntohs
-mdefine_line|#define __constant_ntohs(x) (x)
+mdefine_line|#define __constant_ntohs(x) ((unsigned short int)(x))
 DECL|macro|__constant_htonl
-mdefine_line|#define __constant_htonl(x) (x)
+mdefine_line|#define __constant_htonl(x) ((unsigned long int)(x))
 DECL|macro|__constant_htons
-mdefine_line|#define __constant_htons(x) (x)
+mdefine_line|#define __constant_htons(x) ((unsigned short int)(x))
 macro_line|#ifndef __BIG_ENDIAN
 DECL|macro|__BIG_ENDIAN
 mdefine_line|#define __BIG_ENDIAN 4321
